+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Possible log types:
 - [changed] Require Rust 1.21+
 - [changed] Constructors of `SimpleApi` and `E2eApi` are now private, use the
   `ApiBuilder` instead
+- [added] Add `ApiBuilder::with_custom_endpoint` method
 
 ### v0.7.1 (2017-08-28)